Code Refactored
